+++ src/hooks/useCarousel.ts
@@ -1,0 +1,22 @@
+import { useState } from 'react';
+import { Movie } from '@/models/movie.model';
+
+export function useCarousel(movieList: Movie[]) {
+  const [currentIndex, setCurrentIndex] = useState(0);
+
+  const nextSlide = () => {
+    setCurrentIndex((prevIndex) =>
+      prevIndex === movieList.length - 1 ? 0 : prevIndex + 1,
+    );
+  };
+
+  const prevSlide = () => {
+    setCurrentIndex((prevIndex) =>
+      prevIndex === 0 ? movieList.length - 1 : prevIndex - 1,
+    );
+  };
+
+  const currentMovies = movieList.slice(currentIndex, currentIndex + 3);
+
+  return { currentMovies, nextSlide, prevSlide };
+}
